+++ data/index.js
@@ -1,11 +1,17 @@
 const makeForwardsResponse = require('./make_forwards_response');
 const makeInvoice = require('./make_invoice');
+const makePayViaRoutesResponse = require('./make_pay_via_routes_response');
 const makePaymentData = require('./make_payment_data');
 const makePaymentRequest = require('./make_payment_request');
+const makeRoutesResponse = require('./make_routes_response');
+const makeWalletInfoResponse = require('./make_wallet_info_response');
 
 module.exports = {
   makeForwardsResponse,
   makeInvoice,
+  makePayViaRoutesResponse,
   makePaymentData,
   makePaymentRequest,
+  makeRoutesResponse,
+  makeWalletInfoResponse,
 };