+++ lnd/make_lnd.js
@@ -1,7 +1,11 @@
+const makeBlocksSubscription = require('./make_blocks_subscription');
 const makeForwardsResponse = require('./../data/make_forwards_response');
 const makeInvoice = require('./../data/make_invoice');
 const makeInvoiceSubscription = require('./make_invoice_subscription');
 const makePaySubscription = require('./make_pay_subscription');
+const {makePayViaRoutesResponse} = require('./../data');
+const {makeRoutesResponse} = require('./../data');
+const {makeWalletInfoResponse} = require('./../data');
 
 const bufAsHex = buf => buf.toString('hex');
 
@@ -10,15 +14,21 @@ const bufAsHex = buf => buf.toString('hex');
   {
     getForwards: <Override Get Forwards Response Function>
     getInvoice: <Override Get Invoice Response Function>
+    payViaRoutes: <Override Pay Via Routes Function>
     subscribeToInvoice: <Override Subscribe to Invoice Emitter>
     subscribeToPay: <Override Subscribe to Pay Emitter>
   }
 
   @returns
   {
+    chain: {
+      registerBlockEpochNtfn: ({}) => {}
+    }
     default: {
       addInvoice: ({}, cbk) => {}
+      getInfo: ({}, cbk) => {}
       lookupInvoice: ({}, cbk) => {}
+      queryRoutes: ({}, cbk) => {}
     }
     invoices: {
       subscribeSingleInvoice: ({}) => {}
@@ -30,6 +40,11 @@ const bufAsHex = buf => buf.toString('hex');
 */
 module.exports = overrides => {
   return {
+    chain: {
+      registerBlockEpochNtfn: ({}) => {
+        return makeBlocksSubscription({})
+      },
+    },
     default: {
       addInvoice: ({}, cbk) => {
         return cbk(null, makeInvoice({}));
@@ -42,6 +57,9 @@ module.exports = overrides => {
 
         return cbk(null, makeForwardsResponse({offset: args.index_offset}));
       },
+      getInfo: (args, cbk) => {
+        return cbk(null, makeWalletInfoResponse({}));
+      },
       lookupInvoice: (args, cbk) => {
         // Exit early when there is a custom invoice response
         if (!!overrides.getInvoice) {
@@ -49,6 +67,9 @@ module.exports = overrides => {
         }
 
         return cbk(null, makeInvoice({}));
+      },
+      queryRoutes: ({}, cbk) => {
+        return cbk(null, makeRoutesResponse({}));
       },
     },
     invoices: {
@@ -62,13 +83,21 @@ module.exports = overrides => {
       },
     },
     router: {
-      sendPaymentV2: args => {
+      sendPaymentV2: ({}) => {
         // Exit early when there is a custom payment subscription
         if (!!overrides.subscribeToPay) {
           return overrides.subscribeToPay;
         }
 
         return makePaySubscription({});
+      },
+      sendToRoute: (args, cbk) => {
+        // Exit early when overriding the send to route
+        if (!!overrides.payViaRoutes) {
+          return overrides.payViaRoutes(args, cbk);
+        }
+
+        return cbk(null, makePayViaRoutesResponse({}));
       },
     },
   };